Removed a few exception throws and added caching
Removed "throws Exception" in FavoriteDAOStub's "save" as well as FavoriteServiceStub's delete method because they weren't being used. I also saw that one of the labelled miscellaneous tasks in the project was adding caching, so I went ahead and added some of the annotations for that even though the database isn't set up.